fix bug RTIMER_CLOCK_DIFF(a,b)

+++ utils/inc/rtimer.h
@@ -59,7 +59,7 @@
 #ifndef RTIMER_CLOCK_DIFF
 //typedef unsigned short rtimer_clock_t;
 typedef uint32_t rtimer_clock_t;
-#define RTIMER_CLOCK_DIFF(a,b)     ((signed short)((a)-(b)))
+#define RTIMER_CLOCK_DIFF(a,b)     ((int32_t)((a)-(b)))
 #endif /* RTIMER_CLOCK_DIFF */
 
 #define RTIMER_CLOCK_LT(a, b)      (RTIMER_CLOCK_DIFF((a),(b)) < 0)
